make the drop table migration safer

+++ db/migrate/20150710164755_drop_servers.rb
@@ -1,6 +1,6 @@
 class DropServers < ActiveRecord::Migration
   def change
-    drop_table :affections
-    drop_table :servers
+    drop_table :affections if ActiveRecord::Base.connection.table_exists? :affections
+    drop_table :servers if ActiveRecord::Base.connection.table_exists? :servers
   end
 end
